repository 接口通用性改动 #8051

+++ src/backend/ci/core/stream/biz-stream/src/main/kotlin/com/tencent/devops/stream/trigger/git/service/TGitApiService.kt
@@ -277,7 +277,7 @@ class TGitApiService @Autowired constructor(
             apiErrorCode = ErrorCodeEnum.GET_GIT_FILE_TREE_ERROR
         ) {
             client.get(ServiceGitResource::class).getGitFileTree(
-                gitProjectId = gitProjectId.toLong(),
+                gitProjectId = gitProjectId,
                 path = path ?: "",
                 token = cred.toToken(),
                 ref = ref,
